+++ src/main/java/com/iaali/ota_users_service/service/impl/ProfileServiceImpl.java
@@ -121,7 +121,9 @@ public class ProfileServiceImpl implements ProfileService {
 
     @Override
     public void hardDelete(Long id) {
-        repository.findById(id).orElseThrow(() -> new GlobalException(id, ErrorEnum.NOT_FOUND_ID));
+        if (!repository.existsById(id)) {
+            throw new GlobalException(id, ErrorEnum.NOT_FOUND_ID);
+        }
 
         repository.deleteById(id);
         // Associated user is automatically deleted due to CascadeType.REMOVE
